requirements.txt
numpy==1.24.3
Pillow==9.5.0
progress==1.6
pyserial==3.5
six==1.16.0
tabulate==0.9.0
